implement scroll to location card on a map marker click
